feat: static-rule/flat/SB3 agents + env parity validation
Static GGR rule inside the Gymnasium env reproduces
src.backtest.simulate_pair_returns exactly (Milestone 1 gate).
Observations switched to float64 so threshold decisions match the
reference simulator's precision.

+++ tests/test_rl_episodes.py
@@ -56,7 +56,7 @@ def test_episode_shapes_consistent(episodes):
     for ep in episodes:
         assert len(ep.p1) == len(ep.p2) == len(ep.dates) == ep.length
         assert ep.features.shape == (ep.length, len(STATIC_FEATURES))
-        assert ep.features.dtype == np.float32
+        assert ep.features.dtype == np.float64
         assert not np.isnan(ep.features).any()
         assert (ep.p1 > 0).all() and (ep.p2 > 0).all()
         assert ep.locked_sigma > 0
